+++ rswag-specs/lib/generators/rspec/swagger_generator.rb
@@ -6,13 +6,14 @@ require 'rails/generators'
 module Rspec
   class SwaggerGenerator < ::Rails::Generators::NamedBase
     source_root File.expand_path('templates', __dir__)
+    class_option :spec_path, type: :string, default: 'requests'
 
     def setup
       @routes = Rswag::RouteParser.new(controller_path).routes
     end
 
     def create_spec_file
-      template 'spec.rb', File.join('spec', 'requests', "#{controller_path}_spec.rb")
+      template 'spec.rb', File.join('spec', options['spec_path'], "#{controller_path}_spec.rb")
     end
 
     private