+++ rswag-specs/lib/rswag/specs/example_group_helpers.rb
@@ -13,9 +13,9 @@ module Rswag
       end
 
       [:get, :post, :patch, :put, :delete, :head, :options, :trace].each do |verb|
-        define_method(verb) do |summary, &block|
-          api_metadata = { operation: { verb: verb, summary: summary } }
-          describe(verb, api_metadata, &block)
+        define_method(verb) do |summary, **metadata, &block|
+          api_metadata = { operation: { verb: verb, summary: summary } }.deep_merge(metadata)
+          describe(verb, **api_metadata, &block)
         end
       end
 
@@ -55,17 +55,16 @@ module Rswag
         end
       end
 
-
-      def request_body_example(value:, summary: nil, name: nil) 
-        if metadata.key?(:operation) 
+      def request_body_example(value:, summary: nil, name: nil)
+        if metadata.key?(:operation)
           metadata[:operation][:request_examples] ||= []
-          example = { value: value } 
-          example[:summary] = summary if summary 
+          example = { value: value }
+          example[:summary] = summary if summary
           # We need the examples to have a unique name for a set of examples, so just make the name the length if one isn't provided.
           example[:name] = name || metadata[:operation][:request_examples].length()
           metadata[:operation][:request_examples] << example
-        end 
-      end 
+        end
+      end
 
       def response(code, description, metadata = {}, &block)
         metadata[:response] = { code: code, description: description }
@@ -115,14 +114,27 @@ module Rswag
         )
       end
 
-      def run_test!(&block)
+      #
+      # Perform request and assert response matches swagger definitions
+      #
+      # @param description [String] description of the test
+      # @param args [Array] arguments to pass to the `it` method
+      # @param options [Hash] options to pass to the `it` method
+      # @param &block [Proc] you can make additional assertions within that block
+      # @return [void]
+      def run_test!(description = nil, *args, **options, &block)
+        # rswag metadata value defaults to true
+        options[:rswag] = true unless options.key?(:rswag)
+
+        description ||= "returns a #{metadata[:response][:code]} response"
+
         if RSPEC_VERSION < 3
           ActiveSupport::Deprecation.warn('Rswag::Specs: WARNING: Support for RSpec 2.X will be dropped in v3.0')
           before do
             submit_request(example.metadata)
           end
 
-          it "returns a #{metadata[:response][:code]} response", rswag: true do
+          it description, *args, **options do
             assert_response_matches_metadata(metadata)
             block.call(response) if block_given?
           end
@@ -131,7 +143,7 @@ module Rswag
             submit_request(example.metadata)
           end
 
-          it "returns a #{metadata[:response][:code]} response", rswag: true do |example|
+          it description, *args, **options do |example|
             assert_response_matches_metadata(example.metadata, &block)
             example.instance_exec(response, &block) if block_given?
           end