+++ rswag-specs/spec/generators/rspec/swagger_generator_spec.rb
@@ -28,6 +28,16 @@ module Rspec
       end
     end
 
+    it 'generates spec file for a controller in a defined directory' do
+      allow_any_instance_of(Rswag::RouteParser).to receive(:routes).and_return(fake_routes)
+      run_generator %w(Posts::CommentsController --spec_path=integration)
+
+      assert_file('spec/integration/posts/comments_spec.rb') do |content|
+        assert_match(/parameter name: 'post_id', in: :path, type: :string/, content)
+        assert_match(/patch\('update_comments comment'\)/, content)
+      end
+    end
+
     private
 
     def fake_routes