+++ rswag-specs/spec/rswag/specs/request_factory_spec.rb
@@ -31,15 +31,15 @@ module Rswag
         end
 
         context "'path' parameters" do
-          context 'when `name` parameter key is required, but not defined within example group' do
-            before do
-              metadata[:path_item][:template] = '/blogs/{blog_id}/comments/{id}'
-              metadata[:operation][:parameters] = [
-                { name: 'blog_id', in: :path, type: :number },
-                { name: 'id', in: :path, type: :number }
-              ]
-            end
+          before do
+            metadata[:path_item][:template] = '/blogs/{blog_id}/comments/{id}'
+            metadata[:operation][:parameters] = [
+              { name: 'blog_id', in: :path, type: :number },
+              { name: 'id', in: :path, type: :number }
+            ]
+          end
 
+          context 'when `name` parameter key is required, but not defined within example group' do
             it "explicitly warns user about missing parameter, instead of giving generic error" do
               expect { request[:path] }.not_to raise_error(/undefined method/)
               expect { request[:path] }.not_to raise_error(/is not available from within an example/)
@@ -49,17 +49,27 @@ module Rswag
 
           context 'when `name` is defined' do
             before do
-              metadata[:path_item][:template] = '/blogs/{blog_id}/comments/{id}'
-              metadata[:operation][:parameters] = [
-                { name: 'blog_id', in: :path, type: :number },
-                { name: 'id', in: :path, type: :number }
-              ]
               allow(example).to receive(:blog_id).and_return(1)
               allow(example).to receive(:id).and_return(2)
             end
 
             it 'builds the path from example values' do
               expect(request[:path]).to eq('/blogs/1/comments/2')
+            end
+
+            context 'when `getter is defined`' do
+              before do
+                metadata[:operation][:parameters] = [
+                  { name: 'blog_id', in: :path, type: :number },
+                  { name: 'id', in: :path, type: :number, getter: :param_id }
+                ]
+
+                allow(example).to receive(:param_id).and_return(123)
+              end
+
+              it 'builds the path using getter method' do
+                expect(request[:path]).to eq('/blogs/1/comments/123')
+              end
             end
           end
         end
@@ -77,48 +87,88 @@ module Rswag
           it 'builds the query string from example values' do
             expect(request[:path]).to eq('/blogs?q1=foo&q2=bar')
           end
+
+          context 'when `getter is defined`' do
+            before do
+              metadata[:operation][:parameters] << {
+                name: 'status', in: :query, type: :string, getter: :q3_status
+              }
+
+              allow(example).to receive(:status).and_return(nil)
+              allow(example).to receive(:q3_status).and_return(123)
+            end
+
+            it 'builds the query string using getter method' do
+              expect(request[:path]).to eq('/blogs?q1=foo&q2=bar&status=123')
+            end
+          end
         end
 
         context "'query' parameters of type 'array'" do
           before do
             metadata[:operation][:parameters] = [
-              { name: 'things', in: :query, type: :array, collectionFormat: collection_format }
+              { name: 'things', in: :query, type: :array, collectionFormat: collection_format },
+              { name: 'numbers', in: :query, type: :array, collectionFormat: collection_format, getter: :magic_numbers },
             ]
             allow(example).to receive(:things).and_return(['foo', 'bar'])
+            allow(example).to receive(:magic_numbers).and_return([0, 1])
+            expect(example).not_to receive(:numbers)
           end
 
           context 'collectionFormat = csv' do
             let(:collection_format) { :csv }
             it 'formats as comma separated values' do
-              expect(request[:path]).to eq('/blogs?things=foo,bar')
+              expect(request[:path]).to eq('/blogs?things=foo,bar&numbers=0,1')
             end
           end
 
           context 'collectionFormat = ssv' do
             let(:collection_format) { :ssv }
             it 'formats as space separated values' do
-              expect(request[:path]).to eq('/blogs?things=foo bar')
+              expect(request[:path]).to eq('/blogs?things=foo bar&numbers=0 1')
             end
           end
 
           context 'collectionFormat = tsv' do
             let(:collection_format) { :tsv }
             it 'formats as tab separated values' do
-              expect(request[:path]).to eq('/blogs?things=foo\tbar')
+              expect(request[:path]).to eq('/blogs?things=foo\tbar&numbers=0\t1')
             end
           end
 
           context 'collectionFormat = pipes' do
             let(:collection_format) { :pipes }
             it 'formats as pipe separated values' do
-              expect(request[:path]).to eq('/blogs?things=foo|bar')
+              expect(request[:path]).to eq('/blogs?things=foo|bar&numbers=0|1')
             end
           end
 
           context 'collectionFormat = multi' do
             let(:collection_format) { :multi }
             it 'formats as multiple parameter instances' do
-              expect(request[:path]).to eq('/blogs?things=foo&things=bar')
+              expect(request[:path]).to eq('/blogs?things=foo&things=bar&numbers=0&numbers=1')
+            end
+          end
+        end
+
+        context "'query' parameter of format 'datetime'" do
+          let(:date_time) { DateTime.new(2001, 2, 3, 4, 5, 6, '-7').to_s  }
+
+          before do
+            metadata[:operation][:parameters] = [
+              { name: 'date_time', in: :query, type: :string, format: :datetime, }
+            ]
+            allow(example).to receive(:date_time).and_return(date_time)
+          end
+
+          it 'formats the datetime properly' do
+            expect(request[:path]).to eq('/blogs?date_time=2001-02-03T04%3A05%3A06-07%3A00')
+          end
+
+          context "iso8601 format" do
+            let(:date_time) { DateTime.new(2001, 2, 3, 4, 5, 6, '-7').iso8601 }
+            it 'is also formatted properly' do
+              expect(request[:path]).to eq('/blogs?date_time=2001-02-03T04%3A05%3A06-07%3A00')
             end
           end
         end
@@ -271,12 +321,16 @@ module Rswag
 
         context "'header' parameters" do
           before do
-            metadata[:operation][:parameters] = [{ name: 'Api-Key', in: :header, type: :string }]
+            metadata[:operation][:parameters] = [
+              { name: 'Api-Key', in: :header, type: :string },
+              { name: 'Token', getter: :token_param, in: :header, type: :string }
+            ]
             allow(example).to receive(:'Api-Key').and_return('foobar')
+            allow(example).to receive(:'token_param').and_return('my_token')
           end
 
           it 'adds names and example values to headers' do
-            expect(request[:headers]).to eq({ 'Api-Key' => 'foobar' })
+            expect(request[:headers]).to eq({ 'Api-Key' => 'foobar', 'Token' => 'my_token' })
           end
         end
 
@@ -589,8 +643,11 @@ module Rswag
           context 'openapi 3.0' do
             before do
               swagger_doc[:servers] = [{
-                :url => "https://{defaultHost}",
+                :url => "{protocol}://{defaultHost}",
                 :variables => {
+                  :protocol => {
+                    :default => :https
+                  },
                   :defaultHost => {
                     :default => "www.example.com"
                   }